fix: update application with React Router.
Updates entire application with React Router.
