+++ src/components/Banner/banner.js
@@ -1,10 +1,9 @@
 import "./banner.css";
 import React from "react";
 import BannerImage from "./BannerImage";
+import { Link } from "react-router-dom";
 
 const Banner = (props) => {
-  const { setPage } = props;
-
   /*-- XL- 1450 * 720 lg- 950 * 350 md- 992 * 576 sm- 768 * 576 --*/
 
   return (
@@ -14,10 +13,9 @@ const Banner = (props) => {
       <div className="banner-text-wrapper top-center-container ">
         <h1 className="banner-title">Your one stop shop</h1>
         <h2>At a discount price</h2>
-        <button className="btn btn-dark" onClick={() => setPage("all")}>
+        <Link className="btn btn-dark" to="/react-e-store/category/">
           View All Items
-          <i className="fal fa-chevron-right"></i>
-        </button>
+        </Link>
       </div>
     </div>
   );
